Alterado o atributo compensation para float

+++ include/nanvix/pm.h
@@ -201,7 +201,7 @@
     	int counter;             /**< Remaining quantum.      */
     	int priority;            /**< Process priorities.     */
     	int tickets;             /**< Process tickets.        */
-    	int compensation;        /**< Tickets compensation.   */ 
+    	float compensation;      /**< Tickets compensation.   */ 
     	int nice;                /**< Nice for scheduling.    */
     	unsigned alarm;          /**< Alarm.                  */
 		struct process *next;    /**< Next process in a list. */
